wip, lots of typescripting
