+++ packages/data-management/src/HierarchicalTrackSelectorDrawerWidget/components/HierarchicalTrackSelector.js
@@ -130,7 +130,9 @@ function HierarchicalTrackSelector({ model }) {
 
   const { assemblyNames } = model
   const assemblyName = assemblyNames[assemblyIdx]
-  if (!assemblyName) return null
+  if (!assemblyName) {
+    return null
+  }
   const filterError =
     model.trackConfigurations(assemblyName) > 0 &&
     model.trackConfigurations(assemblyName).filter(filter).length === 0
