+++ packages/jbrowse-desktop/src/declare.d.ts
@@ -1,3 +1,1 @@
-declare module '@gmod/jbrowse-core/configuration'
-declare module '@gmod/jbrowse-core/configuration/configurationSchema'
 declare module '@gmod/jbrowse-core/util/tracks'
